+++ cmake/dependencies_core.cmake
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: Copyright (c) 2021-2024, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-FileCopyrightText: Copyright (c) 2021-2025, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -18,6 +18,10 @@ list(APPEND CMAKE_MESSAGE_CONTEXT "dep_core")
 # cccl -- get an explicit cccl build, matx tries to pull a tag that doesn't exist.
 # =========
 morpheus_utils_configure_cccl()
+
+# indicators
+# ==========
+morpheus_utils_configure_indicators()
 
 # matx
 # ====