+++ tests/morpheus/test_config.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-# SPDX-FileCopyrightText: Copyright (c) 2022-2024, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-FileCopyrightText: Copyright (c) 2022-2025, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -159,6 +159,7 @@ def test_warning_model_batch_size_less_than_pipeline_batch_size(caplog: pytest.L
     config.pipeline_batch_size = 256
     with caplog.at_level(logging.WARNING):
         config.model_max_batch_size = 257
+        config.freeze()
         assert len(caplog.records) == 1
         import re
         assert re.match(".*pipeline_batch_size < model_max_batch_size.*", caplog.records[0].message) is not None
@@ -169,6 +170,7 @@ def test_warning_pipeline_batch_size_less_than_model_batch_size(caplog: pytest.L
     config.model_max_batch_size = 8
     with caplog.at_level(logging.WARNING):
         config.pipeline_batch_size = 7
+        config.freeze()
         assert len(caplog.records) == 1
         import re
         assert re.match(".*pipeline_batch_size < model_max_batch_size.*", caplog.records[0].message) is not None